refactor: combine query running capabilities
